refactor(web): group review control tests
