+++ apps/web/src/screens/review/tests/controls/ReviewScreenControlTestSupport.tsx
@@ -2,7 +2,7 @@ import { act } from "react";
 import {
   clickElement,
   dispatchKeydown,
-} from "../testSupport/ReviewScreenTestSupport";
+} from "../../testSupport/ReviewScreenTestSupport";
 
 export async function flushReviewScreenPromises(): Promise<void> {
   await act(async () => {
@@ -58,3 +58,5 @@ export function getActiveReviewFilterOption(activeOptionOwner: HTMLElement): HTM
 
   return activeOption;
 }
+
+
